move utils to own directory and started logger
