+++ deno.lock
@@ -9,15 +9,22 @@
     "jsr:@cloudydeno/kubernetes-client@0.7.3": "0.7.3",
     "jsr:@cloudydeno/kubernetes-client@0.7.5": "0.7.5",
     "jsr:@cloudydeno/stream-observables@1": "1.4.1",
+    "jsr:@deno-lib/logger@^1.1.9": "1.1.9",
+    "jsr:@deno-library/logger@*": "1.2.0",
+    "jsr:@deno-library/logger@^1.2.0": "1.2.0",
     "jsr:@henrygd/semaphore@^0.0.2": "0.0.2",
     "jsr:@std/bytes@^1.0.2": "1.0.6",
+    "jsr:@std/fmt@1.0.3": "1.0.3",
     "jsr:@std/fmt@~1.0.2": "1.0.8",
+    "jsr:@std/fs@^1.0.6": "1.0.19",
+    "jsr:@std/internal@^1.0.10": "1.0.10",
     "jsr:@std/internal@^1.0.9": "1.0.10",
-    "jsr:@std/path@1": "1.1.1",
+    "jsr:@std/path@1": "1.1.2",
     "jsr:@std/path@1.0.6": "1.0.6",
-    "jsr:@std/streams@1": "1.0.10",
+    "jsr:@std/path@^1.1.1": "1.1.2",
+    "jsr:@std/streams@1": "1.0.11",
     "jsr:@std/streams@1.0.5": "1.0.5",
-    "jsr:@std/text@~1.0.7": "1.0.15",
+    "jsr:@std/text@~1.0.7": "1.0.16",
     "jsr:@std/yaml@1": "1.0.9",
     "jsr:@std/yaml@1.0.5": "1.0.5",
     "jsr:@std/yaml@1.0.9": "1.0.9",
@@ -30,7 +37,7 @@
         "jsr:@cliffy/flags",
         "jsr:@cliffy/internal",
         "jsr:@cliffy/table",
-        "jsr:@std/fmt",
+        "jsr:@std/fmt@~1.0.2",
         "jsr:@std/text"
       ]
     },
@@ -46,7 +53,7 @@
     "@cliffy/table@1.0.0-rc.8": {
       "integrity": "8bbcdc2ba5e0061b4b13810a24e6f5c6ab19c09f0cce9eb691ccd76c7c6c9db5",
       "dependencies": [
-        "jsr:@std/fmt"
+        "jsr:@std/fmt@~1.0.2"
       ]
     },
     "@cloudydeno/kubernetes-apis@0.5.2": {
@@ -75,14 +82,38 @@
     "@cloudydeno/stream-observables@1.4.1": {
       "integrity": "555aa3d786fcfd066d096fc943ea47afc688618e95146e615705e981008f57c0"
     },
+    "@deno-lib/logger@1.1.9": {
+      "integrity": "b03b8044d5ed655a73261aebae0d9f998823f8776f7c7e172e8f4b44f2688ab2",
+      "dependencies": [
+        "jsr:@std/fmt@1.0.3",
+        "jsr:@std/fs"
+      ]
+    },
+    "@deno-library/logger@1.2.0": {
+      "integrity": "c6f0371cb2d08c0a7b82d30d494fac5d25f115695c63a3a295c4e35a0e6669a0",
+      "dependencies": [
+        "jsr:@std/fmt@1.0.3",
+        "jsr:@std/fs"
+      ]
+    },
     "@henrygd/semaphore@0.0.2": {
       "integrity": "cdf678250474b9445648d4fb8d8c636fe313f54893a50de616bf8d3e14f3b51e"
     },
     "@std/bytes@1.0.6": {
       "integrity": "f6ac6adbd8ccd99314045f5703e23af0a68d7f7e58364b47d2c7f408aeb5820a"
     },
+    "@std/fmt@1.0.3": {
+      "integrity": "97765c16aa32245ff4e2204ecf7d8562496a3cb8592340a80e7e554e0bb9149f"
+    },
     "@std/fmt@1.0.8": {
       "integrity": "71e1fc498787e4434d213647a6e43e794af4fd393ef8f52062246e06f7e372b7"
+    },
+    "@std/fs@1.0.19": {
+      "integrity": "051968c2b1eae4d2ea9f79a08a3845740ef6af10356aff43d3e2ef11ed09fb06",
+      "dependencies": [
+        "jsr:@std/internal@^1.0.9",
+        "jsr:@std/path@^1.1.1"
+      ]
     },
     "@std/internal@1.0.10": {
       "integrity": "e3be62ce42cab0e177c27698e5d9800122f67b766a0bea6ca4867886cbde8cf7"
@@ -93,7 +124,13 @@
     "@std/path@1.1.1": {
       "integrity": "fe00026bd3a7e6a27f73709b83c607798be40e20c81dde655ce34052fd82ec76",
       "dependencies": [
-        "jsr:@std/internal"
+        "jsr:@std/internal@^1.0.9"
+      ]
+    },
+    "@std/path@1.1.2": {
+      "integrity": "c0b13b97dfe06546d5e16bf3966b1cadf92e1cc83e56ba5476ad8b498d9e3038",
+      "dependencies": [
+        "jsr:@std/internal@^1.0.10"
       ]
     },
     "@std/streams@1.0.5": {
@@ -105,8 +142,14 @@
     "@std/streams@1.0.10": {
       "integrity": "75c0b1431873cd0d8b3d679015220204d36d3c7420d93b60acfc379eb0dc30af"
     },
+    "@std/streams@1.0.11": {
+      "integrity": "db583d27e28d133f389f1eec318cffdf4998305e5134c1d4b1c56b361cee6018"
+    },
     "@std/text@1.0.15": {
       "integrity": "91f5cc1e12779a3d95f1be34e763f9c28a75a078b7360e6fcaef0d8d9b1e3e7f"
+    },
+    "@std/text@1.0.16": {
+      "integrity": "ddb9853b75119a2473857d691cf1ec02ad90793a2e8b4a4ac49d7354281a0cf8"
     },
     "@std/yaml@1.0.5": {
       "integrity": "71ba3d334305ee2149391931508b2c293a8490f94a337eef3a09cade1a2a2742"
@@ -131,6 +174,8 @@
       "jsr:@cliffy/command@1.0.0-rc.8",
       "jsr:@cloudydeno/kubernetes-apis@0.5.2",
       "jsr:@cloudydeno/kubernetes-client@0.7.5",
+      "jsr:@deno-lib/logger@^1.1.9",
+      "jsr:@deno-library/logger@^1.2.0",
       "jsr:@henrygd/semaphore@^0.0.2",
       "jsr:@std/yaml@1.0.9"
     ]
