+++ src/logic/mod.ts
@@ -1,5 +1,4 @@
 import { Codefresh } from './codefresh.ts';
 import { K8s } from './k8s.ts';
-import { Utils } from './utils.ts';
 
-export { Codefresh, K8s, Utils };
+export { Codefresh, K8s };
