+++ src/utils/logger.ts
@@ -1,0 +1,3 @@
+import { Logger } from 'jsr:@deno-library/logger';
+
+export const logger = new Logger();
