+++ src/utils/mod.ts
@@ -1,0 +1,4 @@
+import { logger } from './logger.ts';
+import { Utils } from './utils.ts';
+
+export { logger, Utils };
